refactor: 구독 버튼 교체 컴포넌트 재사용성 향상
